Adding option to disable liveness analysis
Dyninst's liveness analysis tries to predict when it is safe to use a
general purpose register without saving it first. This was seen to cause
issues (particularly in codes compiled with -O2) where GP registers are
used across function call boundaries. The "-a", "save all", option
disables this analysis and forces dyninst to save all GP registers.
Note: this can increase the overhead per instruction due to additional
push/pops.